% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CySA-package.R
\docType{package}
\name{CySA-package}
\alias{CySA}
\alias{CySA-package}
\title{CySA: Interactive Cluster Selector for Cytometry Data}
\description{
Provides an interactive Shiny application for selecting and visualizing
clusters from flow cytometry data stored in \code{SingleCellExperiment}
objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baj12/CySA}
  \item Report bugs at \url{https://github.com/baj12/CySA/issues}
}

}
\author{
\strong{Maintainer}: Bernd Jagla \email{bernd.jagla@pasteur.fr} (\href{https://orcid.org/0000-0002-7696-0484}{ORCID})

Authors:
\itemize{
  \item Bernd Jagla \email{bernd.jagla@pasteur.fr} (\href{https://orcid.org/0000-0002-7696-0484}{ORCID})
}

}
\keyword{internal}
