% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CySA-package.R
\name{CySA_default_cluster_cols}
\alias{CySA_default_cluster_cols}
\title{Default cluster color palette}
\usage{
CySA_default_cluster_cols()
}
\value{
A character vector of hex colors.
}
\description{
Returns the default 20-color palette used by CySA for cluster
visualizations.
}
\examples{
CySA_default_cluster_cols()

}
