% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_clusterSelector.R
\name{.buildSOMRasterBox}
\alias{.buildSOMRasterBox}
\title{Build SOM Raster UI}
\usage{
.buildSOMRasterBox()
}
\value{
A \code{shinydashboardPlus::box} UI object.
}
\description{
Raster heatmap of SOM node marker expression plus a small interactive plot
for selecting nodes from the SOM grid.
}
\keyword{internal}
