% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClusterSelector.R
\name{.buildSOMRasterData}
\alias{.buildSOMRasterData}
\title{Build SOM Raster Data}
\usage{
.buildSOMRasterData(sce, somCodesName = "SOM_codes")
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} with SOM codes in
\code{metadata(sce)[[somCodesName]]} and optional \code{SOM_stats}.}
}
\value{
A \code{data.frame}.
}
\description{
Builds a data frame representing the SOM grid, including the SOM codes and
per-node cell counts.
}
\keyword{internal}
