% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{.buildUpSetBox}
\alias{.buildUpSetBox}
\title{Build UpSet Plot UI}
\usage{
.buildUpSetBox(outputList)
}
\arguments{
\item{outputList}{Named list of cluster groupings.}
}
\value{
A \code{shinydashboardPlus::box} UI object.
}
\description{
Build UpSet Plot UI
}
\keyword{internal}
