% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-reactive-helpers.R
\name{.buildZoomObservers}
\alias{.buildZoomObservers}
\title{Build Zoom Observers for SOM Data Plots}
\usage{
.buildZoomObservers(nPlots, input, zoomFunc, verbose)
}
\arguments{
\item{nPlots}{Number of 2D SOM plots.}

\item{input}{Shiny input object.}

\item{zoomFunc}{Function that applies zoom state to \code{dimSelection}.}

\item{verbose}{Logical indicating whether to print debug messages.}
}
\description{
Registers one \code{plotly_relayout} observer per 2D SOM plot.
}
\keyword{internal}
