% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{.computeRelativeCounts}
\alias{.computeRelativeCounts}
\title{Compute Relative Cell Counts}
\usage{
.computeRelativeCounts(
  clusterPatientTable,
  rs,
  relativeToCol,
  expInfo,
  outputList
)
}
\arguments{
\item{clusterPatientTable}{Table of sample by cluster counts.}

\item{rs}{Selected cluster ids.}

\item{relativeToCol}{Reference column or group name.}

\item{expInfo}{Sample-level metadata data.frame.}

\item{outputList}{Named list of cluster groupings.}
}
\value{
Numeric vector of percentages.
}
\description{
Computes the percentage of selected cells relative to a reference column,
another cluster group, or the total population.
}
\keyword{internal}
