% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClusterSelector.R
\name{.defaultDList}
\alias{.defaultDList}
\title{Default Marker Pairs for 2D Plots}
\usage{
.defaultDList(sce)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment}.}
}
\value{
A list of character pairs.
}
\description{
Builds a default list of marker pairs from a \code{SingleCellExperiment}
object.
}
\keyword{internal}
