% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClusterSelector.R
\name{.isDendrogramLeaf}
\alias{.isDendrogramLeaf}
\title{Test if an Object is a Dendrogram Leaf}
\usage{
.isDendrogramLeaf(x)
}
\arguments{
\item{x}{A dendrogram object.}
}
\value{
Logical.
}
\description{
Test if an Object is a Dendrogram Leaf
}
\keyword{internal}
