% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClusterSelector.R
\name{.nDendrogramLeaves}
\alias{.nDendrogramLeaves}
\title{Count Leaves in a Dendrogram}
\usage{
.nDendrogramLeaves(x)
}
\arguments{
\item{x}{A dendrogram object.}
}
\value{
Integer.
}
\description{
Count Leaves in a Dendrogram
}
\keyword{internal}
