% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{.updateOutputListInputs}
\alias{.updateOutputListInputs}
\title{Synchronize Select Input Choices with outputList}
\usage{
.updateOutputListInputs(session, input, outputList, metaD)
}
\arguments{
\item{session}{Shiny session object.}

\item{input}{Shiny input object.}

\item{outputList}{Named list of cluster groupings.}

\item{metaD}{\code{metadata(sce)} list.}
}
\description{
Updates the UI select inputs that depend on the names of \code{outputList}.
}
\keyword{internal}
