% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterBJ.R
\name{plotScatterBJ}
\alias{plotScatterBJ}
\title{Plot Scatter for Bioconductor Experiment}
\usage{
plotScatterBJ(
  x,
  chs,
  gate = NULL,
  color_by = NULL,
  facet_by = NULL,
  bins = 100,
  assay = "exprs",
  label = c("target", "channel", "both"),
  zeros = FALSE,
  k_pal = NULL,
  rowNames = NULL
)
}
\arguments{
\item{x}{A Bioconductor Experiment object.}

\item{chs}{A character vector specifying the variables to be plotted on the x and y axes.}

\item{gate}{Optional gate object.}

\item{color_by}{A character specifying the variable used for color coding points (optional).}

\item{facet_by}{A character specifying the variable used for faceting the plot (optional).}

\item{bins}{Number of bins for the 2D histogram (default is 100).}

\item{assay}{The assay to use for plotting (default is "exprs").}

\item{label}{Label for points on the plot ("target", "channel", or "both").}

\item{zeros}{Logical, whether to exclude rows with all zero values (default is FALSE).}

\item{k_pal}{Color palette for plotting.}

\item{rowNames}{Optional character vector to specify row names for the plot.}
}
\value{
A ggplot2 scatter plot.
}
\description{
This function generates a scatter plot for a Bioconductor Experiment object.
}
\examples{
sce <- CySA_example_sce()
plotScatterBJ(sce, chs = c("marker1", "marker2"))

}
