% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CySA-package.R
\name{CySA_default_cluster_cols}
\alias{CySA_default_cluster_cols}
\title{CySA: Interactive Cluster Selector for Cytometry Data}
\usage{
CySA_default_cluster_cols()
}
\description{
Provides an interactive Shiny application for selecting and visualizing
clusters from flow cytometry data stored in \code{SingleCellExperiment}
objects.
}
\keyword{internal}
