% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{.buildClusterSelectorSidebar}
\alias{.buildClusterSelectorSidebar}
\title{Build clusterSelector Sidebar UI}
\usage{
.buildClusterSelectorSidebar(sce, dList, outputList, nPlots)
}
\arguments{
\item{sce}{Full \code{SingleCellExperiment} for marker choices.}

\item{dList}{List of marker pairs used to label the plot index selector.}

\item{outputList}{Named list of cluster groupings.}

\item{nPlots}{Number of 2D SOM plots.}
}
\value{
A \code{shinydashboard::dashboardSidebar} UI object.
}
\description{
Build clusterSelector Sidebar UI
}
\keyword{internal}
