% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{.buildFlowSOMPieBox}
\alias{.buildFlowSOMPieBox}
\title{Build FlowSOM Marker Pies UI}
\usage{
.buildFlowSOMPieBox()
}
\value{
A \code{shinydashboardPlus::box} UI object.
}
\description{
Build FlowSOM Marker Pies UI
}
\keyword{internal}
