% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{.buildFlowSOMStarsBox}
\alias{.buildFlowSOMStarsBox}
\title{Build FlowSOM Stars UI}
\usage{
.buildFlowSOMStarsBox(fsom)
}
\arguments{
\item{fsom}{Optional \code{FlowSOM} object. If \code{NULL}, returns
\code{NULL}.}
}
\value{
A \code{shinydashboardPlus::box} UI object or \code{NULL}.
}
\description{
Build FlowSOM Stars UI
}
\keyword{internal}
