% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{.buildSOM2DPlotsBox}
\alias{.buildSOM2DPlotsBox}
\title{Build SOM 2D Plots UI}
\usage{
.buildSOM2DPlotsBox(nPlots, colsUsed, outputList)
}
\arguments{
\item{nPlots}{Number of 2D SOM plots to display.}

\item{colsUsed}{Character vector of SOM column names used for dim red.}

\item{outputList}{Named list of cluster groupings.}
}
\value{
A \code{shinydashboardPlus::box} UI object.
}
\description{
Build SOM 2D Plots UI
}
\keyword{internal}
