% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-reactive-helpers.R
\name{.buildSOMDataObservers}
\alias{.buildSOMDataObservers}
\title{Build SOM Data Selection Observers}
\usage{
.buildSOMDataObservers(
  nPlots,
  input,
  output,
  rsUsed,
  rsUsed_d,
  activePlot,
  inputSelect,
  verbose
)
}
\arguments{
\item{nPlots}{Number of 2D SOM plots.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{rsUsed}{Reactive value holding the currently selected nodes.}

\item{activePlot}{Reactive value tracking the active plot index.}

\item{inputSelect}{Function that maps event data to new selected ids.}

\item{verbose}{Logical indicating whether to print debug messages.}
}
\description{
Loops over the 2D SOM plot outputs and registers selection observers.
}
\keyword{internal}
