% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{.buildSOMRasterBox}
\alias{.buildSOMRasterBox}
\title{Build SOM Raster UI}
\usage{
.buildSOMRasterBox()
}
\value{
A \code{shinydashboardPlus::box} UI object.
}
\description{
Build SOM Raster UI
}
\keyword{internal}
